Refactor mock data
