+++ product-service/tsconfig.json
@@ -10,7 +10,7 @@
     "target": "ES2020",
     "outDir": "lib"
   },
-  "include": ["src/**/*.ts", "serverless.ts"],
+  "include": ["src/**/*.ts", "serverless.ts", "tests/**/*.ts"],
   "exclude": [
     "node_modules/**/*",
     ".serverless/**/*",
